git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@159736 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: db1b5c55010a074d0cf84c0dd44395df7f990571

+++ Online/Online/DIM/src/did/did.c
@@ -1083,8 +1083,8 @@ int get_nodes();
 		XmListAddItem(id,create_str(node),i+1);
 	}
 	*/
-	set_something(id,XmNlistItemCount,(void*)i);
-	set_something(id,XmNlistVisibleItemCount,(void*)(i < 8) ? i : 8);
+	set_something(id,XmNlistItemCount,(void*)(long)i);
+	set_something(id,XmNlistVisibleItemCount,(void*)(long)(i < 8) ? i : 8);
 }	
 
 void get_server_service()
@@ -2902,8 +2902,8 @@ void delete_str();
 	}
 	free(service_list);
 	
-	set_something(id,XmNlistItemCount,(void*)i);
-	set_something(id,XmNlistVisibleItemCount,(void*)(i < 20) ? i : 20);
+	set_something(id,XmNlistItemCount,(void*)(long)i);
+	set_something(id,XmNlistVisibleItemCount,(void*)(long)(i < 20) ? i : 20);
 }
 
 void show_clients(SERVER **servp_ptr, char *buffer, int *size)
